requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
ucimlrepo
pyaml
mlflow
fastapi
python-multipart
uvicorn
black

#-e .